requirements.txt
openai>=1.93.0
python-dotenv>=1.0.0
streamlit>=1.46.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
beautifulsoup4>=4.12.0
youtube-transcript-api>=0.6.1
langchain>=0.1.0
langchain-openai>=0.3.0
faiss-cpu>=1.7.4
chromadb>=0.4.15
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0
jinja2>=3.1.2
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0
playwright>=1.53.0 